Add files via upload
This project uses the Stern-Brocot tree to implement a binary tree structure to represent rational numbers. Using L for left and R for right, the program can convert between corresponding paths in the tree and rational numbers (in the form p/q). It also determines and outputs the total amount of time that each conversion made during the session takes to execute. This project shows how number theory and tree structures can be mapped efficiently, which is especially helpful for issues with data structures and algorithms.